LogOutButton moved to own component
